Luka/Income/requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.28.2
flask-swagger-ui==4.11.1
flask-cors

